refactor(pocket3d): move the character controller into pocket3d-bsp
collide.rs (TraceWorld + GoldSrc movement: friction/accelerate/air-cap/
step/jump) relocates from the wgpu crate to the no_std map substrate, so
the PSP game runtime runs the exact same movement source as desktop.
pocket3d::collide stays as a re-export; the prelude is unchanged.
